+++ src/Rocks/Extensions/IParameterSymbolExtensions.cs
@@ -6,5 +6,6 @@ internal static class IParameterSymbolExtensions
 {
 	internal static bool RequiresForcedNullableAnnotation(this IParameterSymbol self) =>
 		self.HasExplicitDefaultValue && self.ExplicitDefaultValue is null &&
+			self.Type.TypeKind != TypeKind.TypeParameter &&
 			!self.Type.IsValueType && self.NullableAnnotation != NullableAnnotation.Annotated;
 }